Select permanentes segun su registro, stock en 0 apoyo visual, precio…

+++ database/migrations/2022_11_27_214720_create_tb_usuarios_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('nombre');
             $table->string('apellidoP');
             $table->string('apellidoM');
-            $table->integer('celular');
+            $table->bigInteger('celular');
             $table->date('fechaIngreso');
             $table->string('email');
             $table->string('contraseña');
